cleaning up group caller
